#40 Add Spotify connection in danceMove activity. Co-authored-by: b4rdha <gusahmeba@student.gu.se>
