+++ Application/src/main/AndroidManifest.xml
