+++ Application/src/main/java/com/example/android/dancecar/DanceMoves.java
@@ -17,6 +17,11 @@ import android.widget.TextView;
 import android.widget.Toast;
 
 
+import com.spotify.android.appremote.api.ConnectionParams;
+import com.spotify.android.appremote.api.Connector;
+import com.spotify.android.appremote.api.SpotifyAppRemote;
+import com.spotify.protocol.types.Track;
+
 import org.eclipse.paho.client.mqttv3.IMqttActionListener;
 import org.eclipse.paho.client.mqttv3.IMqttDeliveryToken;
 import org.eclipse.paho.client.mqttv3.IMqttToken;
@@ -36,6 +41,9 @@ public class DanceMoves extends AppCompatActivity {
     LinearLayout lLayout;
     LinearLayout rLayout;
     CheckBox checkBox;
+    private static final String CLIENT_ID = "764ef5ad07284dd499fcb8bb5604bc26";
+    private static final String REDIRECT_URI = "dancing-car-app://callback";
+    private SpotifyAppRemote mSpotifyAppRemote;
 
 
     private  Button createChor;
@@ -198,6 +206,49 @@ public class DanceMoves extends AppCompatActivity {
         Intent intent = new Intent(DanceMoves.this, NewMoves.class);
         startActivity(intent);
     }
+
+    public void goToSpotify(View view){
+        ConnectionParams connectionParams = new ConnectionParams.Builder(CLIENT_ID).setRedirectUri(REDIRECT_URI).showAuthView(true).build();
+
+        Log.d(TAG, "onStart: blabla : " + connectionParams);
+        SpotifyAppRemote.connect(getApplicationContext(), connectionParams, new Connector.ConnectionListener() {
+
+            public void onConnected(SpotifyAppRemote spotifyAppRemote) {
+                mSpotifyAppRemote = spotifyAppRemote;
+                Log.d("MainActivity", "Connected! Yay!");
+                // Now you can start interacting with App Remote
+                connected();
+
+            }
+
+            public void onFailure(Throwable throwable) {
+                Log.e("MyActivity", throwable.getMessage(), throwable);
+
+                // Something went wrong when attempting to connect! Handle errors here
+            }
+        });
+    }
+
+    protected void onStop() {
+        super.onStop();
+        SpotifyAppRemote.disconnect(mSpotifyAppRemote);
+    }
+
+    private void connected() {
+        // Play a playlist
+        mSpotifyAppRemote.getPlayerApi().play("spotify:playlist:37i9dQZF1DX2sUQwD7tbmL");
+
+        // Subscribe to PlayerState
+        mSpotifyAppRemote.getPlayerApi()
+                .subscribeToPlayerState()
+                .setEventCallback(playerState -> {
+                    final Track track = playerState.track;
+                    if (track != null) {
+                        Log.d("MainActivity", track.name + " by " + track.artist.name);
+                    }
+                });
+    }
+
 
     View.OnClickListener checkBoxDance(final CheckBox checkBox, final ChorMoves chor){
         return new View.OnClickListener() {
