+++ Application/src/main/res/layout/activity_dance_moves.xml
@@ -296,13 +296,31 @@
         android:layout_marginEnd="180dp"
         android:layout_marginRight="180dp"
         android:layout_marginBottom="216dp"
-        android:onClick="makeCarDance"
         android:background="@color/cardview_dark_background"
+        android:onClick="makeCarDance"
         app:layout_constraintBottom_toBottomOf="parent"
         app:layout_constraintEnd_toEndOf="parent"
         app:layout_constraintStart_toStartOf="parent"
         app:layout_constraintTop_toTopOf="parent"
         app:srcCompat="@android:drawable/ic_media_play" />
+
+    <Button
+        android:id="@+id/button6"
+        android:layout_width="105dp"
+        android:layout_height="31dp"
+        android:layout_marginStart="37dp"
+        android:layout_marginLeft="37dp"
+        android:layout_marginTop="406dp"
+        android:layout_marginEnd="269dp"
+        android:layout_marginRight="269dp"
+        android:layout_marginBottom="294dp"
+        android:text="Spotify"
+        android:textSize="10sp"
+        android:onClick="goToSpotify"
+        app:layout_constraintBottom_toBottomOf="parent"
+        app:layout_constraintEnd_toEndOf="parent"
+        app:layout_constraintStart_toStartOf="parent"
+        app:layout_constraintTop_toTopOf="parent" />
 
 
 </androidx.constraintlayout.widget.ConstraintLayout>
