+++ spotify-app/build.gradle
@@ -1,2 +1,4 @@
 configurations.maybeCreate("default")
 artifacts.add("default", file('spotify-app-remote-release-0.7.1.aar'))
+
+
